fix: major river cell insertion bugs + file detection fix
- RiverGeometry.coordinates: float[][] → double[][] (GeoJSON coords are doubles)
- River.ControlPoints: List<float[]> → List<double[]> to match
- RiverImageGenerator: cast double coords to float for GeoToPixel call
- MajorRiverInserter: skip engulfed cells that contain a burg (would destroy settlement)
- MajorRiverInserter: update burg.Cell back-reference after split/clipped land cell creation
- ModManager.FindLatestInputs: fix false-positive rivers detection — a file containing
  "cells" is always the cells GeoJSON even if the map name contains "rivers"
- ImageUtility.DrawCells/DrawCellsWithNeighborLines: add fileName parameter (default unchanged)
- ConversionManager: draw post-river-insertion debug images (1_cells_post_river.png,
  1_cells_neighbors_post_river.png) so cell splits are visible without waiting for baronies

+++ Converter/Lemur/Deserialization/AzgaarData.cs
@@ -281,7 +281,7 @@ namespace Converter.Lemur.Deserialization
 
     public record RiverGeometry(
         string type,        // "LineString"
-        float[][] coordinates  // Array of [x, y] coordinate pairs
+        double[][] coordinates  // Array of [x, y] coordinate pairs
     );
 
     public record RiverFeatureProperties(
